requirements.txt
# Core dependencies
langchain
langchain-community
langchain-openai
openai
faiss-cpu
python-dotenv
tiktoken
aiohttp
pandas
numpy
matplotlib
seaborn
scikit-learn
requests
python-dotenv
uuid